change: refactor process.py
